PROBLEM:
Prevent double loading of the page when including passphrase login.

CAUSE:
Passphrase login needs to make the browser's password manager recognize the secret key, and there is a hack in that template to use a form going to an iframe.IframeTransport

SOLUTION:
Make sure that hack doesn't actually load the current page.